刪除用戶功能

+++ client/src/pages/Homepage/PatchProfile/PatchProfile.js
@@ -26,21 +26,50 @@ export default function PatchProfile({ currentUser, setCurrentUser }) {
 
   // 處理修改個人資料
   const handlePatchProfile = async () => {
-    try {
-      let response = await AuthService.patchProfile(currentUser.user._id, username, email); 
-      window.alert("修改成功。您現在將被導向到首頁");
-      localStorage.setItem("user", JSON.stringify(response.data)); 
-      setCurrentUser(AuthService.getCurrentUser()); 
-      navigate("/"); // 導航至首頁
-    } catch (e) {
-      setMessage(e.response.data); 
-    }
+    const confirmResult = window.confirm("您確定要修改資料嗎？");
+    if (confirmResult) {
+      try {
+        let response = await AuthService.patchProfile(currentUser.user._id, username, email);
+        window.alert("修改成功。您現在將被導向到首頁");
+        localStorage.setItem("user", JSON.stringify(response.data));
+        setCurrentUser(AuthService.getCurrentUser());
+        navigate("/"); // 導航至首頁
+      } catch (e) {
+        setMessage(e.response.data);
+      };
+    };
+  };
+
+  // 處裡刪除個人資料
+  const handleDeleteProfile = async () => {
+    const confirmResult = window.confirm("您確定要刪除帳號嗎？");
+    if (confirmResult) {
+      try {
+        await AuthService.deleteUser(currentUser.user._id);
+        window.alert("刪除帳號成功。您現在將被導向到登入畫面");
+        setCurrentUser(null);
+        navigate("/firstEnroll");
+      } catch (e) {
+        if (e.response && e.response.data) {
+          window.alert(e.response.data.message);
+        } else {
+          window.alert("發生了一些錯誤，請稍後再試。");
+        };
+      };
+    };
   };
 
   return (
     <div className="flex justify-center w-full h-full items-center">
       {/* 修改個人資料表單 */}
-      <div className="rounded-lg border border-blue-400 bg-blue-100 dark:text-black p-5 shadow-xl dark:bg-gray-800 dark:border-gray-700 w-3/5 min-w-52">
+      <div className="relative rounded-lg border border-blue-400 bg-blue-100 dark:text-black p-5 shadow-xl dark:bg-gray-800 dark:border-gray-700 w-3/5 min-w-52">
+        {/* 返回鍵 */}
+        <div onClick={() => navigate("/")} className="absolute right-1 top-1 h-5 w-5 cursor-pointer bg-white">
+          <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24">
+            <path fill="none" d="M0 0h24v24H0z" />
+            <path d="M12 10.586l-4.293-4.293-1.414 1.414L10.586 12l-4.293 4.293 1.414 1.414L12 13.414l4.293 4.293 1.414-1.414L13.414 12l4.293-4.293-1.414-1.414L12 10.586z" />
+          </svg>
+        </div>
         {/* 用戶名稱輸入框 */}
         <div className="mb-5">
           <label for="username"><div className="mb-2 ml-2 text-sm dark:text-white">用戶名稱</div></label>
@@ -68,8 +97,8 @@ export default function PatchProfile({ currentUser, setCurrentUser }) {
           <span className="flex items-center text-sm md:text-base justify-center text-red-600 md:justify-start">{message && message}</span>
           {/* 返回按鈕和確認按鈕 */}
           <div className="mt-3 flex justify-center md:mt-0 md:justify-end">
-            <button onClick={() => navigate("/")} className="text-nowrap rounded border border-gray-300 bg-gray-200 px-2 py-1 text-sm font-bold text-white shadow-sm md:px-5 md:py-1 md:text-base dark:bg-gray-800 dark:text-gray-400 dark:border-gray-700 hover:bg-gray-600">返回</button>
-            <button onClick={handlePatchProfile} className="text-nowrap rounded border border-blue-300 bg-blue-500 px-2 py-1 text-sm font-bold text-white shadow-sm md:px-5 md:py-1 md:text-base dark:bg-gray-900 dark:text-blue-500 dark:border-gray-700 hover:bg-blue-500 hover:text-black">確認</button>
+            <button onClick={handleDeleteProfile} className="text-nowrap rounded border border-gray-300 bg-gray-200 px-2 py-1 text-sm font-bold text-white shadow-sm md:px-5 md:py-1 md:text-base dark:bg-gray-800 dark:text-gray-400 dark:border-gray-700 hover:bg-gray-600 hover:text-red-400">刪除帳號</button>
+            <button onClick={handlePatchProfile} className="text-nowrap rounded border border-blue-300 bg-blue-500 px-2 py-1 text-sm font-bold text-white shadow-sm md:px-5 md:py-1 md:text-base dark:bg-gray-900 dark:text-blue-500 dark:border-gray-700 hover:bg-blue-500 hover:text-black">確認修改</button>
           </div>
         </div>
       </div>
